Added enemy that follows. Had to switch everything from int to float

+++ SimpleGame/src/com/relegatedthings/simplegame/BasicEnemy.java
@@ -8,7 +8,7 @@ public class BasicEnemy extends GameObject {
 
 	private Handler handler;
 	
-	public BasicEnemy(int x, int y, ID id, Handler handler) {
+	public BasicEnemy(float x, float y, ID id, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		
@@ -18,7 +18,7 @@ public class BasicEnemy extends GameObject {
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int) x, (int) y, 16, 16);
 	}
 	
 	public void tick() {
@@ -32,12 +32,12 @@ public class BasicEnemy extends GameObject {
 			velX *= -1;
 		}
 		
-		handler.addObject(new Trail(x, y, ID.Trail, Color.red, 16, 16, .02f, handler));
+		handler.addObject(new Trail((int) x, (int) y, ID.Trail, Color.red, 16, 16, .02f, handler));
 	}
 
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int) x, (int) y, 16, 16);
 		
 	}
